requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.11.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
numpy==1.26.4
sentence-transformers>=2.0.0
